Let major version of oldest installed kernel determine baseline
Based on the confusion I've read in recent bug reports, this is an
attempt to improve the default behavior of mainline-gtk with respect
to what kernel versions it chooses to display by default.

The idea here is to let the oldest installed kernel version help
determine what range of versions to display by default. For example,
if the oldest installed version is 5.4, then it will by default
display all versions starting from 5.0 and newer. If you want more
versions displayed, you can still use the old option of showing N
previous major versions.

If the oldest version cannot be determined by querying the packaging
manager, the version of the current running kernel will be used.