Call `X.init_threads()` at startup
This is required as, in future commits, we are going to update GTK
widgets in other threads than the main one. Hence, using the GDK
method `threads_add_idle_full()`. It however requires Xlib mutexes
to be initialized, i.e. calling `X.init_threads()`.

This is necessary for now but we should be able to remove it in the
future. Indeed, it seems that more recent version of Gtk remove this
dependency.