Introduce `FileDownloader` a new utility to download file
This utility aims to be part of a replacement system for
`DownloadManager`. In fact, the latter suffer from multiple flaws as
its unnecessary complexity, dependence over an external tool and its
design forcing the user to busy-wait.

`FileDownloader` is only based on `gio`, so does not require an external
tool. It also is simple as it doesn't have to manage outputs of its
fork, supports progress reports and cancellation. And finally, is easily
embeddable in another class to create a download pool.

This patch also adds a simple usage of this utility: download the main
`index.html` file.