Download changelogs lazily
`CHANGES` files used to be downloaded when filling the cache on the
first start. However, these files are not actually required by the
application's standard workflow. This patch, using the newly introduced
`FileDownloader`, allows to postpone the download of the changelog until
its usage.

As a consequence of reducing the amount of files to download, a cold
start now takes 50% less times on my computer. Dropping from 36 to only
18 seconds. There is still room for improvement, but it remains a good
start :^).