Log programming errors instead of popping a dialog
These buttons should be disabled if calling them doesn't make sense. If
a user achieve to call it anyway, it is a programming error. Those
don't need to be reported to the end user but only to the developer.

Hence, this path makes the application log errors with `return_if_fail`
instead of displaying a dialog.